+++ index.js
@@ -6,7 +6,17 @@ const RNBackgroundGeolocation = NativeModules.BackgroundGeolocation;
 function emptyFn() {}
 
 var BackgroundGeolocation = {
-  events: ['location', 'stationary', 'error'],
+  events: [
+    'location',
+    'stationary',
+    'start',
+    'stop',
+    'error',
+    'mode_change',
+    'permissions_denied',
+    'foreground',
+    'background'
+  ],
 
   provider: {
     ANDROID_DISTANCE_FILTER_PROVIDER: 0,
@@ -31,22 +41,26 @@ var BackgroundGeolocation = {
     RNBackgroundGeolocation.configure(config, successFn, errorFn);
   },
 
-  start: function(successFn, errorFn) {
-    successFn = successFn || emptyFn;
-    errorFn = errorFn || emptyFn;
-    RNBackgroundGeolocation.start(successFn, errorFn);
+  start: function() {
+    RNBackgroundGeolocation.start();
   },
 
-  stop: function(successFn, errorFn) {
-    successFn = successFn || emptyFn;
-    errorFn = errorFn || emptyFn;
-    RNBackgroundGeolocation.stop(successFn, errorFn);
+  stop: function() {
+    RNBackgroundGeolocation.stop();
   },
 
+  // @deprecated
   isLocationEnabled: function(successFn, errorFn) {
+    console.log('[WARN]: this method is deprecated. Use checkStatus instead.');
     successFn = successFn || emptyFn;
     errorFn = errorFn || emptyFn;
     RNBackgroundGeolocation.isLocationEnabled(successFn, errorFn);
+  },
+
+  checkStatus: function(successFn, errorFn) {
+    successFn = successFn || emptyFn;
+    errorFn = errorFn || emptyFn;
+    RNBackgroundGeolocation.checkStatus(successFn, errorFn);
   },
 
   showAppSettings: function() {
@@ -57,24 +71,12 @@ var BackgroundGeolocation = {
     RNBackgroundGeolocation.showLocationSettings();
   },
 
-  watchLocationMode: function(successFn, errorFn) {
-    successFn = successFn || emptyFn;
-    errorFn = errorFn || emptyFn;
-    RNBackgroundGeolocation.watchLocationMode(successFn, errorFn);
-  },
-
-  stopWatchingLocationMode: function(successFn, errorFn) {
-    successFn = successFn || emptyFn;
-    errorFn = errorFn || emptyFn;
-    RNBackgroundGeolocation.stopWatchingLocationMode(successFn, errorFn);
-  },
-
   getLocations: function(successFn, errorFn) {
     successFn = successFn || emptyFn;
     errorFn = errorFn || emptyFn;
     RNBackgroundGeolocation.getLocations(successFn, errorFn);
   },
-/*
+
   getValidLocations: function(successFn, errorFn) {
     successFn = successFn || emptyFn;
     errorFn = errorFn || emptyFn;
@@ -92,7 +94,7 @@ var BackgroundGeolocation = {
     errorFn = errorFn || emptyFn;
     RNBackgroundGeolocation.deleteAllLocations(successFn, errorFn);
   },
-*/
+
   switchMode: function(modeId, successFn, errorFn) {
     successFn = successFn || emptyFn;
     errorFn = errorFn || emptyFn;
@@ -104,7 +106,7 @@ var BackgroundGeolocation = {
     errorFn = errorFn || emptyFn;
     RNBackgroundGeolocation.getConfig(successFn, errorFn);
   },
-  
+
   getLogEntries: function(limit, successFn, errorFn) {
     successFn = successFn || emptyFn;
     errorFn = errorFn || emptyFn;
@@ -120,6 +122,15 @@ var BackgroundGeolocation = {
     }
 
     return DeviceEventEmitter.addListener(event, callbackFn);
+  },
+
+  removeAllListeners: function(event) {
+    if (this.events.indexOf(event) < 0) {
+      console.log('[WARN] RNBackgroundGeolocation: removeAllListeners for unknown event "' + event + '"');
+      return false;
+    }
+
+    return DeviceEventEmitter.removeAllListeners(event);
   }
 };